+++ src/main/java/test/crudboard/controller/MainController.java
@@ -7,32 +7,25 @@ import lombok.extern.slf4j.Slf4j;
 import org.springframework.data.domain.Page;
 import org.springframework.data.domain.PageRequest;
 import org.springframework.data.domain.Sort;
-import org.springframework.security.access.prepost.PreAuthorize;
 import org.springframework.security.core.annotation.AuthenticationPrincipal;
 import org.springframework.stereotype.Controller;
 import org.springframework.ui.Model;
 import org.springframework.web.bind.annotation.*;
 import test.crudboard.domain.entity.post.dto.PostHeaderDto;
 import test.crudboard.domain.entity.post.dto.SearchRequestDto;
-
-import test.crudboard.domain.entity.user.dto.UserInfoDto;
-import test.crudboard.domain.entity.user.dto.UserJoinDto;
 import test.crudboard.security.provider.JwtUserDetails;
-import test.crudboard.repository.JpaUserRepository;
 import test.crudboard.service.post.PostRequestService;
 import test.crudboard.service.post.PostService;
-import test.crudboard.service.UserService;
 
 @Controller
 @Slf4j
 @RequiredArgsConstructor
 public class MainController {
-    private final UserService userService;
     private final PostService postService;
-    private final JpaUserRepository userRepository;
     private final PostRequestService postRequestService;
 
     private final int PAGE_SIZE = 10;
+
     @GetMapping(value = {"","/","/{page}"})
     public String home(@AuthenticationPrincipal JwtUserDetails userDetails, Model model,
                        @RequestParam(value = "isrecommend", required = false, defaultValue = "false") boolean isRecommend,
@@ -67,32 +60,6 @@ public class MainController {
         int currentBlock = (int) Math.ceil((double) page / size);
         return (currentBlock - 1) * size + 1;
     }
-
-    @GetMapping("/login")
-    public String login() {
-        return "login";
-    }
-
-    @GetMapping("/join")
-    public String join(){
-        return "join";
-    }
-
-    @PostMapping("/join")
-    public String joinUser(UserJoinDto userJoinDto){
-        userService.save(userJoinDto);
-        return "redirect:/";
-    }
-
-
-    @GetMapping("/mypage/{userId}")
-    @PreAuthorize("isAuthenticated()")
-    public String myPage(@AuthenticationPrincipal JwtUserDetails user, Model model){
-        UserInfoDto userInfo = userService.getUserInfo(user.getUsername());
-        model.addAttribute("userInfo",userInfo);
-        return "user-detail";
-    }
-
 
     @GetMapping("/search")
     public String searchPost(@Valid @ModelAttribute SearchRequestDto search, @AuthenticationPrincipal JwtUserDetails user, Model model){
