+++ src/main/java/test/crudboard/domain/entity/comment/Comment.java
@@ -11,7 +11,6 @@ import org.springframework.data.annotation.CreatedDate;
 import org.springframework.data.jpa.domain.support.AuditingEntityListener;
 import test.crudboard.domain.entity.post.Post;
 import test.crudboard.domain.entity.user.User;
-
 import java.time.LocalDateTime;
 import java.util.ArrayList;
 import java.util.List;
