+++ src/main/java/test/crudboard/domain/entity/post/dto/PostDetailDto.java
@@ -4,8 +4,6 @@ package test.crudboard.domain.entity.post.dto;
 import lombok.AllArgsConstructor;
 import lombok.Getter;
 import test.crudboard.domain.entity.comment.Comment;
-import test.crudboard.domain.entity.comment.dto.PostFooterDto;
-
 import java.util.List;
 
 /**
