+++ src/main/java/test/crudboard/domain/entity/post/dto/PostHeaderDto.java
@@ -4,8 +4,7 @@ import lombok.AllArgsConstructor;
 import lombok.Getter;
 import lombok.NoArgsConstructor;
 import lombok.Setter;
-import org.springframework.data.annotation.Id;
-import org.springframework.data.redis.core.RedisHash;
+
 
 import java.time.LocalDateTime;
