+++ src/main/java/test/crudboard/repository/JpaPostRepository.java
@@ -25,15 +25,7 @@ public interface JpaPostRepository extends JpaRepository<Post, Long> {
             "(select count(l) from Like l where l.post.id = p.id) as like_count, " +
             "(select count(c) from Comment c where c.post.id = p.id) as comment_count, " +
             "p.user.nickname) ";
-/*    @Query(value = "SELECT p.post_id, p.head, p.context, p.view, p.created, " +
-            "(SELECT COUNT(*) FROM likes l WHERE l.post_id = p.post_id) AS like_count, " +
-            "(SELECT COUNT(*) FROM comment c WHERE c.post_id = p.post_id) AS comment_count, " +
-            "u.nickname " +
-            "FROM post p " +
-            "LEFT OUTER JOIN users u ON p.user_id = u.user_id " +
-            "ORDER BY p.post_id DESC",
-            countQuery = "SELECT COUNT(*) FROM post",
-            nativeQuery = true)*/
+
     @Query("select p.id, p.head, p.context, p.view, p.created, "+
             "(select count(l.id) from Like l where l.post.id = p.id),  " +
             "(select count(c.id) from Comment c where c.post.id = p.id), " +
@@ -43,7 +35,6 @@ public interface JpaPostRepository extends JpaRepository<Post, Long> {
             "order by p.id desc limit 20")
     List<Object[]> findPostList(@Param("start")long start, @Param("end") long end);
 
-    //jpa  에서 join문을 만들 때, 명시적으로 join 한 것만 join함. post의 user가 eager로 설정되어도, 지정하지 않으면  지연로딩 처리됨
     @Query(MAIN_TITLE_SQL +
             "from Post p left join p.user u where p.id = :id")
     Optional<PostHeaderDto> findPostHeaderDto(@Param("id") Long id);
@@ -59,33 +50,10 @@ public interface JpaPostRepository extends JpaRepository<Post, Long> {
 
     boolean existsPostByIdAndUserNickname(Long postId, String name);
 
-
-    @Query("select p from Post p left join p.user u order by p.id desc ")
-    List<Post> findAllLimit(Pageable page);
-
     @Query("select count(p.id) from Post p")
     long count();
 
-    @Lock(LockModeType.PESSIMISTIC_WRITE)
-    @Query("SELECT p FROM Post p WHERE p.id = :postId")
-    Optional<Post> findByIdWithLock(@Param("postId") Long postId);
-
-    @Modifying
-    @Query("UPDATE Post p SET p.view = p.view + 1 WHERE p.id = :postId")
-    void incrementViewCount(@Param("postId") Long postId);
-
-
     @Query("select max(p.id) from Post p")
     Long maxPostId();
-
-/*
-
-   @Query("select new test.crudboard.entity.testDto(" +
-            "p.id, p.head, p.view, u.nickname, " +
-            "(select count(l.id) from Like l where l.post.id = p.id), " +
-            "(select count(c.id) from Comment c where c.post.id =p.id))" +
-            "from Post p left join p.user u")
-    Page<testDto> testMethod2(Pageable pageable);
-*/
 
 }
