+++ src/main/java/test/crudboard/repository/JpaUserRepository.java
@@ -5,7 +5,6 @@ import org.springframework.data.jpa.repository.Query;
 import org.springframework.data.repository.query.Param;
 import org.springframework.stereotype.Repository;
 import test.crudboard.domain.entity.user.User;
-
 import java.util.List;
 import java.util.Optional;
